roles/environment/tasks/requirements.txt
google-cloud-storage
kubernetes